+++ js/directive/NewsTitleUnreadCount.js
@@ -11,6 +11,9 @@ app.directive('newsTitleUnreadCount', function ($window) {
     'use strict';
 
     var baseTitle = $window.document.title;
+    var titles = baseTitle.split('-');
+    var appName = titles[0] || 'News';
+    var ownCloudName = titles[1] || 'ownCloud';
 
     return {
         restrict: 'E',
@@ -19,11 +22,9 @@ app.directive('newsTitleUnreadCount', function ($window) {
         },
         link: function (scope, elem, attrs) {
             attrs.$observe('unreadCount', function (value) {
-                var titles = baseTitle.split('-');
-
                 if (value !== '0') {
-                    $window.document.title = titles[0] +
-                        '(' + value + ') - ' + titles[1];
+                    $window.document.title = appName +
+                        '(' + value + ') - ' + ownCloudName;
                 }
             });
         }