+++ js/tests/unit/stubs/OC.js
@@ -12,5 +12,8 @@ window.OC = {
         'use strict';
 
         return '/base' + url;
+    },
+    theme: {
+    	name: 'test-theme'
     }
 };