+++ CHANGELOG.md
@@ -4,6 +4,7 @@ owncloud-news (3.003)
 * **Enhancement**: Less padding right on mobile phone
 * **Enhancement**: Expanded view: remove date on mobile phone
 * **Enhancement**: Compact view: click on title should remove ellipsis
+* **Enhancement**: Respect theme name in tab title
 
 owncloud-news (3.002)
 * **Bugfix**: If a folder is selected, the f and d shortcuts will jump to the previous or next folder subfeeds